requirements.txt
langgraph
langchain
langchain-community
faiss-cpu
fastapi
uvicorn
python-multipart
pypdf
python-docx
openai
pydantic
python-dotenv
